- Add out folder in gitignore. - Add factory for remote channel client. - Update channel_client_protocol_example.c for connecting to specified ip address if specified. - Add some comments in channel_server_protocol_example.c. - Add logging in tor_proxy_example.c, but for some reason it seems to doesn't work anymore. - Replace some size_t return values in socket.h/c by int.

+++ src/unknownecho/network/api/socket/socket.h
@@ -50,9 +50,9 @@ bool ue_socket_destroy(int fd);
 
 bool ue_socket_is_valid(int fd);
 
-size_t ue_socket_send_string(int fd, const char *string, ue_tls_connection *tls);
+int ue_socket_send_string(int fd, const char *string, ue_tls_connection *tls);
 
-size_t ue_socket_send_data(int fd, unsigned char *data, size_t size, ue_tls_connection *tls);
+int ue_socket_send_data(int fd, unsigned char *data, size_t size, ue_tls_connection *tls);
 
 size_t ue_socket_receive_string_sync(int fd, ue_string_builder *sb, bool blocking, ue_tls_connection *tls);
